loadbalancer/reflector: Separate conversions to their own file
In preparation for adding another reflector that uses the same conversion
functions, move them to conversions.go and clean up cell.go to refer to
the K8sReflectorCell only.

Add source argument to the conversion functions.
