+++ pkg/loadbalancer/reflectors/k8s.go
@@ -12,30 +12,24 @@ import (
 	"log/slog"
 	"maps"
 	"net"
-	"net/netip"
 	"slices"
-	"strings"
 	"sync"
 
 	"github.com/cilium/hive/cell"
 	"github.com/cilium/hive/job"
 	"github.com/cilium/statedb"
 	"github.com/cilium/stream"
-	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/sets"
 
 	daemonK8s "github.com/cilium/cilium/daemon/k8s"
-	"github.com/cilium/cilium/pkg/annotation"
 	cmtypes "github.com/cilium/cilium/pkg/clustermesh/types"
 	"github.com/cilium/cilium/pkg/container"
-	"github.com/cilium/cilium/pkg/ip"
 	"github.com/cilium/cilium/pkg/k8s"
 	"github.com/cilium/cilium/pkg/k8s/resource"
 	slim_corev1 "github.com/cilium/cilium/pkg/k8s/slim/k8s/api/core/v1"
 	k8sUtils "github.com/cilium/cilium/pkg/k8s/utils"
-	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/loadbalancer"
 	lbmaps "github.com/cilium/cilium/pkg/loadbalancer/maps"
 	"github.com/cilium/cilium/pkg/loadbalancer/writer"
@@ -57,7 +51,7 @@ const (
 	reflectorWaitTime = 100 * time.Millisecond
 )
 
-// ReflectorCell reflects Kubernetes Service and EndpointSlice objects to the
+// K8sReflectorCell reflects Kubernetes Service and EndpointSlice objects to the
 // load-balancing tables.
 //
 // Note that this implementation uses Resource[*Service] and Resource[*Endpoints],
@@ -65,12 +59,36 @@ const (
 // times. We should instead have a reflector that is built directly on the client-go
 // reflector (k8s.RegisterReflector) and not populate an intermediate cache.Store.
 // But as we're still experimenting it's easier to build on what already exists.
-var ReflectorCell = cell.Module(
-	"reflector",
+var K8sReflectorCell = cell.Module(
+	"k8s-reflector",
 	"Reflects load-balancing state from Kubernetes",
 
+	// Bridge Resource[XYZ] to Observable[Event[XYZ]]. Makes it easier to
+	// test [ReflectorCell].
+	cell.ProvidePrivate(resourcesToStreams),
 	cell.Invoke(RegisterK8sReflector),
 )
+
+type resourceIn struct {
+	cell.In
+	ServicesResource  resource.Resource[*slim_corev1.Service]
+	EndpointsResource resource.Resource[*k8s.Endpoints]
+}
+
+type StreamsOut struct {
+	cell.Out
+	ServicesStream  stream.Observable[resource.Event[*slim_corev1.Service]]
+	EndpointsStream stream.Observable[resource.Event[*k8s.Endpoints]]
+}
+
+// resourcesToStreams extracts the stream.Observable from resource.Resource.
+// This makes the reflector easier to test as its API surface is reduced.
+func resourcesToStreams(in resourceIn) StreamsOut {
+	return StreamsOut{
+		ServicesStream:  in.ServicesResource,
+		EndpointsStream: in.EndpointsResource,
+	}
+}
 
 type reflectorParams struct {
 	cell.In
@@ -202,7 +220,7 @@ func runServiceEndpointsReflector(ctx context.Context, health cell.Health, p ref
 			initServices(txn)
 
 		case resource.Upsert:
-			svc, fes := convertService(p.ExtConfig, p.Log, obj)
+			svc, fes := convertService(p.ExtConfig, p.Log, obj, source.Kubernetes)
 			if svc == nil {
 				return
 			}
@@ -368,352 +386,6 @@ func (rh *reflectorHealth) report() {
 		}
 	}
 	rh.prevError = processingError
-}
-
-var (
-	zeroV4 = cmtypes.MustParseAddrCluster("0.0.0.0")
-	zeroV6 = cmtypes.MustParseAddrCluster("::")
-
-	ingressDummyAddress = cmtypes.MustParseAddrCluster("192.192.192.192")
-	ingressDummyPort    = uint16(9999)
-)
-
-func isIngressDummyEndpoint(l3n4Addr loadbalancer.L3n4Addr) bool {
-	// The ingress and gateway-api controllers (operator/pkg/model/translation/{gateway-api,ingress}) create
-	// a dummy endpoint to force Cilium to reconcile the service. This is no longer required with this new
-	// control-plane, but due to rolling upgrades we cannot remove it immediately. Hence we have the
-	// special handlind here to just ignore this endpoint to avoid populating the tables with unnecessary
-	// data.
-	return l3n4Addr.AddrCluster.Equal(ingressDummyAddress) && l3n4Addr.Port == ingressDummyPort
-}
-
-func isHeadless(svc *slim_corev1.Service) bool {
-	_, headless := svc.Labels[corev1.IsHeadlessService]
-	if strings.ToLower(svc.Spec.ClusterIP) == "none" {
-		headless = true
-	}
-	return headless
-}
-
-func convertService(cfg loadbalancer.ExternalConfig, log *slog.Logger, svc *slim_corev1.Service) (s *loadbalancer.Service, fes []loadbalancer.FrontendParams) {
-	name := loadbalancer.ServiceName{Namespace: svc.Namespace, Name: svc.Name}
-	s = &loadbalancer.Service{
-		Name:                name,
-		Source:              source.Kubernetes,
-		Labels:              labels.Map2Labels(svc.Labels, string(source.Kubernetes)),
-		Selector:            svc.Spec.Selector,
-		Annotations:         svc.Annotations,
-		HealthCheckNodePort: uint16(svc.Spec.HealthCheckNodePort),
-	}
-
-	expType, err := k8s.NewSvcExposureType(svc)
-	if err != nil {
-		log.Warn("Ignoring annotation",
-			logfields.Error, err,
-			logfields.Annotations, annotation.ServiceTypeExposure,
-			logfields.Service, svc.GetName(),
-		)
-	}
-
-	if len(svc.Spec.Ports) > 0 {
-		s.PortNames = map[string]uint16{}
-		for _, port := range svc.Spec.Ports {
-			s.PortNames[port.Name] = uint16(port.Port)
-		}
-	}
-
-	for _, srcRange := range svc.Spec.LoadBalancerSourceRanges {
-		prefix, err := netip.ParsePrefix(srcRange)
-		if err != nil {
-			continue
-		}
-		s.SourceRanges = append(s.SourceRanges, prefix)
-	}
-
-	switch svc.Spec.ExternalTrafficPolicy {
-	case slim_corev1.ServiceExternalTrafficPolicyLocal:
-		s.ExtTrafficPolicy = loadbalancer.SVCTrafficPolicyLocal
-	default:
-		s.ExtTrafficPolicy = loadbalancer.SVCTrafficPolicyCluster
-	}
-
-	if svc.Spec.InternalTrafficPolicy != nil && *svc.Spec.InternalTrafficPolicy == slim_corev1.ServiceInternalTrafficPolicyLocal {
-		s.IntTrafficPolicy = loadbalancer.SVCTrafficPolicyLocal
-	} else {
-		s.IntTrafficPolicy = loadbalancer.SVCTrafficPolicyCluster
-	}
-	// Scopes for NodePort and LoadBalancer. Either just external (policies are the same), or
-	// both external and internal (when one policy is local)
-	scopes := []uint8{loadbalancer.ScopeExternal}
-	twoScopes := (s.ExtTrafficPolicy == loadbalancer.SVCTrafficPolicyLocal) != (s.IntTrafficPolicy == loadbalancer.SVCTrafficPolicyLocal)
-	if twoScopes {
-		scopes = append(scopes, loadbalancer.ScopeInternal)
-	}
-
-	// SessionAffinity
-	if svc.Spec.SessionAffinity == slim_corev1.ServiceAffinityClientIP {
-		s.SessionAffinity = true
-
-		s.SessionAffinityTimeout = time.Duration(int(time.Second) * int(slim_corev1.DefaultClientIPServiceAffinitySeconds))
-		if cfg := svc.Spec.SessionAffinityConfig; cfg != nil && cfg.ClientIP != nil && cfg.ClientIP.TimeoutSeconds != nil && *cfg.ClientIP.TimeoutSeconds != 0 {
-			s.SessionAffinityTimeout = time.Duration(int(time.Second) * int(*cfg.ClientIP.TimeoutSeconds))
-		}
-	}
-
-	if s.IntTrafficPolicy != loadbalancer.SVCTrafficPolicyLocal && isTopologyAware(svc) {
-		s.TrafficDistribution = loadbalancer.TrafficDistributionPreferClose
-	}
-
-	// A service that is annotated as headless has no frontends, even if the service spec contains
-	// ClusterIPs etc.
-	if isHeadless(svc) {
-		return
-	}
-
-	// ClusterIP
-	if expType.CanExpose(slim_corev1.ServiceTypeClusterIP) {
-		var clusterIPs []string
-		if len(svc.Spec.ClusterIPs) > 0 {
-			clusterIPs = slices.Sorted(slices.Values(svc.Spec.ClusterIPs))
-		} else {
-			clusterIPs = []string{svc.Spec.ClusterIP}
-		}
-
-		for _, ip := range clusterIPs {
-			addr, err := cmtypes.ParseAddrCluster(ip)
-			if err != nil {
-				continue
-			}
-
-			if (!cfg.EnableIPv6 && addr.Is6()) || (!cfg.EnableIPv4 && addr.Is4()) {
-				continue
-			}
-
-			for _, port := range svc.Spec.Ports {
-				p := loadbalancer.NewL4Addr(loadbalancer.L4Type(port.Protocol), uint16(port.Port))
-				if p == nil {
-					continue
-				}
-				fe := loadbalancer.FrontendParams{
-					Type:        loadbalancer.SVCTypeClusterIP,
-					PortName:    loadbalancer.FEPortName(port.Name),
-					ServiceName: name,
-					ServicePort: uint16(port.Port),
-				}
-				fe.Address.AddrCluster = addr
-				fe.Address.Scope = loadbalancer.ScopeExternal
-				fe.Address.L4Addr = *p
-				fes = append(fes, fe)
-			}
-		}
-	}
-
-	// NOTE: We always want to do ClusterIP services even when full kube-proxy replacement is disabled.
-	// See https://github.com/cilium/cilium/issues/16197 for context.
-
-	if cfg.KubeProxyReplacement {
-		// NodePort
-		if (svc.Spec.Type == slim_corev1.ServiceTypeNodePort || svc.Spec.Type == slim_corev1.ServiceTypeLoadBalancer) &&
-			expType.CanExpose(slim_corev1.ServiceTypeNodePort) {
-
-			for _, scope := range scopes {
-				for _, family := range getIPFamilies(svc) {
-					if (!cfg.EnableIPv6 && family == slim_corev1.IPv6Protocol) ||
-						(!cfg.EnableIPv4 && family == slim_corev1.IPv4Protocol) {
-						continue
-					}
-					for _, port := range svc.Spec.Ports {
-						if port.NodePort == 0 {
-							continue
-						}
-
-						fe := loadbalancer.FrontendParams{
-							Type:        loadbalancer.SVCTypeNodePort,
-							PortName:    loadbalancer.FEPortName(port.Name),
-							ServiceName: name,
-							ServicePort: uint16(port.Port),
-						}
-
-						switch family {
-						case slim_corev1.IPv4Protocol:
-							fe.Address.AddrCluster = zeroV4
-						case slim_corev1.IPv6Protocol:
-							fe.Address.AddrCluster = zeroV6
-						default:
-							continue
-						}
-
-						p := loadbalancer.NewL4Addr(loadbalancer.L4Type(port.Protocol), uint16(port.NodePort))
-						if p == nil {
-							continue
-						}
-						fe.Address.Scope = scope
-						fe.Address.L4Addr = *p
-						fes = append(fes, fe)
-					}
-				}
-			}
-		}
-
-		// LoadBalancer
-		if svc.Spec.Type == slim_corev1.ServiceTypeLoadBalancer && expType.CanExpose(slim_corev1.ServiceTypeLoadBalancer) {
-			for _, ip := range svc.Status.LoadBalancer.Ingress {
-				if ip.IP == "" {
-					continue
-				}
-
-				addr, err := cmtypes.ParseAddrCluster(ip.IP)
-				if err != nil {
-					continue
-				}
-				if (!cfg.EnableIPv6 && addr.Is6()) || (!cfg.EnableIPv4 && addr.Is4()) {
-					continue
-				}
-
-				for _, scope := range scopes {
-					for _, port := range svc.Spec.Ports {
-						fe := loadbalancer.FrontendParams{
-							Type:        loadbalancer.SVCTypeLoadBalancer,
-							PortName:    loadbalancer.FEPortName(port.Name),
-							ServiceName: name,
-							ServicePort: uint16(port.Port),
-						}
-
-						p := loadbalancer.NewL4Addr(loadbalancer.L4Type(port.Protocol), uint16(port.Port))
-						if p == nil {
-							continue
-						}
-						fe.Address.AddrCluster = addr
-						fe.Address.Scope = scope
-						fe.Address.L4Addr = *p
-						fes = append(fes, fe)
-					}
-				}
-
-			}
-		}
-
-		// ExternalIP
-		for _, ip := range svc.Spec.ExternalIPs {
-			addr, err := cmtypes.ParseAddrCluster(ip)
-			if err != nil {
-				continue
-			}
-			if (!cfg.EnableIPv6 && addr.Is6()) || (!cfg.EnableIPv4 && addr.Is4()) {
-				continue
-			}
-
-			for _, port := range svc.Spec.Ports {
-				fe := loadbalancer.FrontendParams{
-					Type:        loadbalancer.SVCTypeExternalIPs,
-					PortName:    loadbalancer.FEPortName(port.Name),
-					ServiceName: name,
-					ServicePort: uint16(port.Port),
-				}
-
-				p := loadbalancer.NewL4Addr(loadbalancer.L4Type(port.Protocol), uint16(port.Port))
-				if p == nil {
-					continue
-				}
-
-				fe.Address.AddrCluster = addr
-				fe.Address.Scope = loadbalancer.ScopeExternal
-				fe.Address.L4Addr = *p
-				fes = append(fes, fe)
-			}
-		}
-	}
-
-	return
-}
-
-func getIPFamilies(svc *slim_corev1.Service) []slim_corev1.IPFamily {
-	if len(svc.Spec.IPFamilies) == 0 {
-		// No IP families specified, try to deduce them from the cluster IPs
-		if len(svc.Spec.ClusterIP) == 0 || svc.Spec.ClusterIP == slim_corev1.ClusterIPNone {
-			return nil
-		}
-
-		ipv4, ipv6 := false, false
-		if len(svc.Spec.ClusterIPs) > 0 {
-			for _, cip := range svc.Spec.ClusterIPs {
-				if ip.IsIPv6(net.ParseIP(cip)) {
-					ipv6 = true
-				} else {
-					ipv4 = true
-				}
-			}
-		} else {
-			ipv6 = ip.IsIPv6(net.ParseIP(svc.Spec.ClusterIP))
-			ipv4 = !ipv6
-		}
-		families := make([]slim_corev1.IPFamily, 0, 2)
-		if ipv4 {
-			families = append(families, slim_corev1.IPv4Protocol)
-		}
-		if ipv6 {
-			families = append(families, slim_corev1.IPv4Protocol)
-		}
-		return families
-	}
-	return svc.Spec.IPFamilies
-}
-
-func convertEndpoints(cfg loadbalancer.ExternalConfig, ep *k8s.Endpoints) (name loadbalancer.ServiceName, out []loadbalancer.BackendParams) {
-	name = loadbalancer.ServiceName{
-		Name:      ep.ServiceID.Name,
-		Namespace: ep.ServiceID.Namespace,
-	}
-
-	// k8s.Endpoints may have the same backend address multiple times
-	// with a different port name. Collapse them down into single
-	// entry.
-	type entry struct {
-		portNames []string
-		backend   *k8s.Backend
-	}
-	entries := map[loadbalancer.L3n4Addr]entry{}
-
-	for addrCluster, be := range ep.Backends {
-		if (!cfg.EnableIPv6 && addrCluster.Is6()) || (!cfg.EnableIPv4 && addrCluster.Is4()) {
-			continue
-		}
-		for portName, l4Addr := range be.Ports {
-			l3n4Addr := loadbalancer.L3n4Addr{
-				AddrCluster: addrCluster,
-				L4Addr:      *l4Addr,
-			}
-			if isIngressDummyEndpoint(l3n4Addr) {
-				continue
-			}
-			portNames := entries[l3n4Addr].portNames
-			if portName != "" {
-				portNames = append(portNames, portName)
-			}
-			entries[l3n4Addr] = entry{
-				portNames: portNames,
-				backend:   be,
-			}
-		}
-	}
-	for l3n4Addr, entry := range entries {
-
-		state := loadbalancer.BackendStateActive
-		if entry.backend.Terminating {
-			state = loadbalancer.BackendStateTerminating
-		}
-		be := loadbalancer.BackendParams{
-			Address:   l3n4Addr,
-			NodeName:  entry.backend.NodeName,
-			PortNames: entry.portNames,
-			Weight:    loadbalancer.DefaultBackendWeight,
-			Zone:      entry.backend.Zone,
-			ForZones:  entry.backend.HintsForZones,
-			State:     state,
-		}
-		out = append(out, be)
-	}
-	return
 }
 
 // hostPortServiceNamePrefix returns the common prefix for synthetic HostPort services
@@ -945,19 +617,4 @@ func joinObservables[T any](src stream.Observable[T], srcs ...stream.Observable[
 				src.Observe(ctx, emit, comp)
 			}
 		})
-}
-
-func isTopologyAware(svc *slim_corev1.Service) bool {
-	return getAnnotationTopologyAwareHints(svc) ||
-		(svc.Spec.TrafficDistribution != nil &&
-			*svc.Spec.TrafficDistribution == corev1.ServiceTrafficDistributionPreferClose)
-}
-
-func getAnnotationTopologyAwareHints(svc *slim_corev1.Service) bool {
-	// v1.DeprecatedAnnotationTopologyAwareHints has precedence over v1.AnnotationTopologyMode.
-	value, ok := svc.ObjectMeta.Annotations[corev1.DeprecatedAnnotationTopologyAwareHints]
-	if !ok {
-		value = svc.ObjectMeta.Annotations[corev1.AnnotationTopologyMode]
-	}
-	return !(value == "" || value == "disabled" || value == "Disabled")
 }
