+++ pkg/loadbalancer/reflectors/k8s_test.go
@@ -14,6 +14,7 @@ import (
 	slim_discovery_v1 "github.com/cilium/cilium/pkg/k8s/slim/k8s/api/discovery/v1"
 	"github.com/cilium/cilium/pkg/k8s/testutils"
 	"github.com/cilium/cilium/pkg/loadbalancer"
+	"github.com/cilium/cilium/pkg/source"
 )
 
 var (
@@ -32,7 +33,7 @@ func BenchmarkConvertService(b *testing.B) {
 	svc := obj.(*slim_corev1.Service)
 
 	for b.Loop() {
-		convertService(benchmarkExternalConfig, slog.New(slog.DiscardHandler), svc)
+		convertService(benchmarkExternalConfig, slog.New(slog.DiscardHandler), svc, source.Kubernetes)
 	}
 	b.ReportMetric(float64(b.N)/b.Elapsed().Seconds(), "services/sec")
 }
